requirements.txt
streamlit
tiktoken
aiohttp
aiosignal
altair
anyio
async-timeout
attrs
blinker
cachetools
certifi
charset-normalizer
click
dataclasses-json
decorator
exceptiongroup
fastapi
frozenlist
gitdb
GitPython
greenlet
h11
idna
importlib-metadata
Jinja2
jsonschema
langchain
langchainplus-sdk
markdown-it-py
MarkupSafe
marshmallow
marshmallow-enum
mdurl
multidict
mypy-extensions
numexpr
numpy
openai
openapi-schema-pydantic
packaging
pandas
Pillow
protobuf
pyarrow
pydantic
pydeck
Pygments
Pympler
pyrsistent
python-dateutil
python-dotenv
pytz
pytz-deprecation-shim
PyYAML
requests
rich
six
smmap
sniffio
SQLAlchemy
starlette
tenacity
toml
toolz
tornado
tqdm
typing-inspect
typing_extensions
tzdata
tzlocal
urllib3
uvicorn
validators
yarl
zipp
langchain_experimental
python-dotenv
langchain-community
tabulate